Rbac support for extended query tag Apis (#699)
Rbac support for extended query tag Apis

+++ src/Microsoft.Health.Dicom.Core/Features/ExtendedQueryTag/DeleteExtendedQueryTagHandler.cs
@@ -7,15 +7,20 @@ using System.Threading;
 using System.Threading.Tasks;
 using EnsureThat;
 using MediatR;
+using Microsoft.Health.Core.Features.Security.Authorization;
+using Microsoft.Health.Dicom.Core.Exceptions;
+using Microsoft.Health.Dicom.Core.Features.Common;
+using Microsoft.Health.Dicom.Core.Features.Security;
 using Microsoft.Health.Dicom.Core.Messages.ExtendedQueryTag;
 
 namespace Microsoft.Health.Dicom.Core.Features.ExtendedQueryTag
 {
-    public class DeleteExtendedQueryTagHandler : IRequestHandler<DeleteExtendedQueryTagRequest, DeleteExtendedQueryTagResponse>
+    public class DeleteExtendedQueryTagHandler : BaseHandler, IRequestHandler<DeleteExtendedQueryTagRequest, DeleteExtendedQueryTagResponse>
     {
         private readonly IDeleteExtendedQueryTagService _deleteExtendedQueryTagService;
 
-        public DeleteExtendedQueryTagHandler(IDeleteExtendedQueryTagService deleteExtendedQueryTagService)
+        public DeleteExtendedQueryTagHandler(IAuthorizationService<DataActions> authorizationService, IDeleteExtendedQueryTagService deleteExtendedQueryTagService)
+            : base(authorizationService)
         {
             EnsureArg.IsNotNull(deleteExtendedQueryTagService, nameof(deleteExtendedQueryTagService));
             _deleteExtendedQueryTagService = deleteExtendedQueryTagService;
@@ -24,6 +29,12 @@ namespace Microsoft.Health.Dicom.Core.Features.ExtendedQueryTag
         public async Task<DeleteExtendedQueryTagResponse> Handle(DeleteExtendedQueryTagRequest request, CancellationToken cancellationToken)
         {
             EnsureArg.IsNotNull(request, nameof(request));
+
+            if (await AuthorizationService.CheckAccess(DataActions.Delete, cancellationToken) != DataActions.Delete)
+            {
+                throw new UnauthorizedDicomActionException(DataActions.Delete);
+            }
+
             await _deleteExtendedQueryTagService.DeleteExtendedQueryTagAsync(request.TagPath, cancellationToken);
             return new DeleteExtendedQueryTagResponse();
         }
